+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/button/components/button-icon/index.ts
@@ -1,0 +1,1 @@
+export * from './ButtonIcon';
