+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/button/components/button-text/ButtonText.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import { BodySmallSemiBold, BodySmallSemiBoldProps } from '../../typography';
-import { useButtonContext } from '../ButtonContext';
+import { BodySmallSemiBold, BodySmallSemiBoldProps } from '../../../typography';
+import { useButtonContext } from '../../ButtonContext';
 
 export type ButtonTextProps<T extends React.ElementType = 'span'> = BodySmallSemiBoldProps<T>;
 export const StyledButtonText = styled(BodySmallSemiBold)``;
