+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/button/components/button-text/index.ts
@@ -1,0 +1,1 @@
+export * from './ButtonText';
